+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Only listing significant user-visible, not internal code cleanups and minor bug 
 * [QATM-2490] Save service response in environment variable without File parameters
 * [QATM-2494] Fix Permission denied error when we try to copy a file and destination file name is not defined 
 * [QATM-1896] Allow standalone-chrome usage without grid
+* [QATM-2533] Allow local variables in runOnEnv and skipOnEnv tags
 
 ## 0.11.0 (June 11, 2019)